gm_gameworld_test: Major reorg of imports in BUILD + moving globals out of main.cpp to allow testing.
The new test does not work yet, as some initialization code is not run correctly to allow loading the outfits correctly.

Additionally, on macOS 12 with Xcode 14.2 / 14C18, the build currently require -copt=-w to squelch an audiocd-related warning in SDL1.2, plus the drawn content in the window that is opened is completely blank.

- Introduced `cursorutil` library for cursor management with SDL support.
- Updated `CreatureUI` methods to return boolean for success/failure.
- Enhanced `loadGfx` in `Creatures` to handle loading failures.
- Added tests for `GM_Gameworld` to ensure proper functionality.
- Refactored various classes to improve code clarity and error handling.